+++ src/app/items/item-by-id.component.ts
@@ -1,7 +1,6 @@
-import { Component, effect, inject, OnDestroy, signal, viewChild } from '@angular/core';
-import { toSignal } from '@angular/core/rxjs-interop';
+import { Component, inject, OnDestroy, viewChild } from '@angular/core';
 import { RouterLink, RouterLinkActive, RouterOutlet } from '@angular/router';
-import { combineLatest, of, Observable, BehaviorSubject } from 'rxjs';
+import { combineLatest, of, Observable } from 'rxjs';
 import {
   delay,
   distinctUntilChanged,
@@ -11,17 +10,13 @@ import {
 } from 'rxjs/operators';
 import { FetchState, readyState } from 'src/app/utils/state';
 import { CurrentContentService } from 'src/app/services/current-content.service';
-import { breadcrumbServiceTag } from './data-access/get-breadcrumb.service';
 import { ItemData } from './models/item-data';
-import { errorHasTag, errorIsHTTPForbidden, errorIsHTTPNotFound } from 'src/app/utils/errors';
 import { isItemUnavailableError } from './utils/item-unavailable-error';
 import { ItemRouter } from 'src/app/models/routing/item-router';
-import { isATask } from 'src/app/items/models/item-type';
-import { itemInfo } from 'src/app/models/content/item-info';
 import { UserSessionService } from 'src/app/services/user-session.service';
 import { ContentDisplayType, LayoutService } from 'src/app/services/layout.service';
 import { mapStateData } from 'src/app/utils/operators/state';
-import { RawItemRoute, routeWithSelfAttempt } from 'src/app/models/routing/item-route';
+import { RawItemRoute } from 'src/app/models/routing/item-route';
 import { BeforeUnloadComponent } from 'src/app/guards/before-unload-guard';
 import { ItemContentComponent } from './containers/item-content/item-content.component';
 import { PendingChangesComponent } from 'src/app/guards/pending-changes-guard';
@@ -54,6 +49,7 @@ import { ButtonComponent } from 'src/app/ui-components/button/button.component';
 import { LoginWallComponent } from './containers/login-wall/login-wall.component';
 import { RestrictedContentComponent } from './containers/restricted-content/restricted-content.component';
 import { ItemTaskFlowService, selectState } from './item-task-flow.service';
+import { ItemContentSyncService } from './item-content-sync.service';
 
 /**
  * ItemByIdComponent is just a container for detail or edit page but manages the fetching on id change and (un)setting the current content.
@@ -62,7 +58,7 @@ import { ItemTaskFlowService, selectState } from './item-task-flow.service';
   selector: 'alg-item-by-id',
   templateUrl: './item-by-id.component.html',
   styleUrl: './item-by-id.component.scss',
-  providers: [ InitialAnswerDataSource, ItemTabs, ItemTaskFlowService ],
+  providers: [ InitialAnswerDataSource, ItemTabs, ItemTaskFlowService, ItemContentSyncService ],
   imports: [
     ItemHeaderComponent,
     AccessCodeViewComponent,
@@ -91,6 +87,7 @@ export class ItemByIdComponent implements OnDestroy, BeforeUnloadComponent, Pend
   private currentContent = inject(CurrentContentService);
   private itemTabs = inject(ItemTabs);
   private taskFlow = inject(ItemTaskFlowService);
+  private itemContentSync = inject(ItemContentSyncService);
   private userSessionService = inject(UserSessionService);
   private breadcrumbService = inject(ItemBreadcrumbsWithFailoverService);
   private layoutService = inject(LayoutService);
@@ -100,15 +97,13 @@ export class ItemByIdComponent implements OnDestroy, BeforeUnloadComponent, Pend
 
   readonly itemContentComponent = viewChild(ItemContentComponent);
 
-  readonly editorUrl = signal<string | undefined>(undefined);
+  readonly fullFrameContent$ = this.itemContentSync.fullFrameContent$;
+  readonly editorUrl = this.itemContentSync.editorUrl;
 
   /**
    * The general state, either the route error handling state, or if not routing error, the item data
    */
   state$: Observable<FetchState<ItemData>> = this.store.select(selectState).pipe(filter(isNotNull));
-
-  // to prevent looping indefinitely in case of bug in services (wrong path > item without path > fetch path > item with path > wrong path)
-  hasRedirected = false;
 
   showAccessCodeField$ = this.state$.pipe(
     mapStateData(data =>
@@ -120,15 +115,6 @@ export class ItemByIdComponent implements OnDestroy, BeforeUnloadComponent, Pend
 
   currentTab$ = this.itemTabs.currentTab$;
   currentTaskView$ = this.itemTabs.currentTaskView$;
-
-  readonly fullFrameContent$ = new BehaviorSubject<boolean>(false); // feeded by task change (below) and task api (item-content comp)
-
-  private itemDataState = this.store.selectSignal(fromItemContent.selectActiveContentData);
-  private breadcrumbsState = this.store.selectSignal(fromItemContent.selectActiveContentBreadcrumbsState);
-  private fullFrameContentSig = toSignal(this.fullFrameContent$, { initialValue: false });
-  private lastTrackedRouteId: string | undefined;
-  private lastTrackedRouteIdInitialized = false;
-  private lastLayoutConfig: { id: string, display: ContentDisplayType } | undefined;
 
   readonly observedGroup$ = this.store.select(fromObservation.selectObservedGroupInfo);
   readonly isThreadInline$ = combineLatest([
@@ -155,25 +141,6 @@ export class ItemByIdComponent implements OnDestroy, BeforeUnloadComponent, Pend
     this.taskFlow.registerSaveHandler(() =>
       this.itemContentComponent()?.itemDisplayComponent()?.saveAnswerAndState() ?? of(readyState<void>(undefined))
     );
-
-    // Field-initializer subscriptions ran synchronously; effects defer to the first CD flush.
-    this.applyResetOnItemChange();
-    this.applyItemToCurrentContentSync();
-    this.applyBreadcrumbsErrorHandling();
-    this.applyLayoutDisplaySync();
-
-    effect(() => {
-      this.applyResetOnItemChange();
-    });
-    effect(() => {
-      this.applyItemToCurrentContentSync();
-    });
-    effect(() => {
-      this.applyBreadcrumbsErrorHandling();
-    });
-    effect(() => {
-      this.applyLayoutDisplaySync();
-    });
 
     this.breadcrumbService.resultPathStarted$.pipe(takeUntilDestroyed()).subscribe(() => this.currentContent.forceNavMenuReload());
   }
@@ -238,83 +205,6 @@ export class ItemByIdComponent implements OnDestroy, BeforeUnloadComponent, Pend
 
   setTaskView(view: string, route: RawItemRoute): void {
     this.taskFlow.setTaskView(view, route);
-  }
-
-  private applyResetOnItemChange(): void {
-    const state = this.itemDataState();
-    if (state === null) return;
-
-    const itemId = state.data?.route.id;
-    if (!this.lastTrackedRouteIdInitialized) {
-      this.lastTrackedRouteIdInitialized = true;
-      this.lastTrackedRouteId = itemId;
-      this.performItemChangeReset();
-      return;
-    }
-    if (itemId === this.lastTrackedRouteId) return;
-    this.lastTrackedRouteId = itemId;
-    this.performItemChangeReset();
-  }
-
-  // Intentionally writes fullFrameContent$/editorUrl consumed by syncLayoutDisplay and the template.
-  private performItemChangeReset(): void {
-    this.fullFrameContent$.next(false);
-    this.editorUrl.set(undefined);
-    this.itemTabs.itemChanged();
-  }
-
-  private applyItemToCurrentContentSync(): void {
-    const state = this.itemDataState();
-    if (!state?.isReady) return;
-    const data = state.data;
-    this.currentContent.replace(itemInfo({
-      route: routeWithSelfAttempt(data.route, data.currentResult?.attemptId),
-      details: {
-        title: data.item.string.title,
-        type: data.item.type,
-        permissions: data.item.permissions,
-        attemptId: data.currentResult?.attemptId,
-        bestScore: data.item.noScore ? undefined : (data.item.watchedGroup ? data.item.watchedGroup.averageScore : data.item.bestScore),
-        currentScore: data.item.noScore ? undefined :
-          (data.item.watchedGroup ? data.item.watchedGroup.averageScore : data.currentResult?.score),
-        validated: data.item.noScore ? undefined :
-          (data.item.watchedGroup ? data.item.watchedGroup.averageScore === 100 : data.currentResult?.validated),
-        leftNavIcon: data.item.displaySettings.leftNavIcon ?? undefined,
-      },
-    }));
-  }
-
-  private applyBreadcrumbsErrorHandling(): void {
-    const state = this.breadcrumbsState();
-    if (state.isError) this.currentContent.clear();
-
-    if (
-      state.isError &&
-      errorHasTag(state.error, breadcrumbServiceTag) &&
-      (errorIsHTTPForbidden(state.error) || errorIsHTTPNotFound(state.error))
-    ) {
-      if (this.hasRedirected) throw new Error('Too many redirections (unexpected)');
-      this.hasRedirected = true;
-      const route = state.identifier;
-      if (!route) throw new Error('unexpected: the active breadcrumbs state should always have an identifier');
-      const routeWithoutPath = { ...route, path: undefined };
-      this.itemRouter.navigateTo(routeWithoutPath, { navExtras: { replaceUrl: true }, useCurrentObservation: true });
-    }
-    if (state.isReady) this.hasRedirected = false;
-  }
-
-  private applyLayoutDisplaySync(): void {
-    const state = this.itemDataState();
-    if (!state?.isReady) return;
-    const data = state.data;
-    const fullFrame = this.fullFrameContentSig();
-    const display = fullFrame
-      ? ContentDisplayType.ShowFullFrame
-      : (isATask(data.item) ? ContentDisplayType.Show : ContentDisplayType.Default);
-    const id = data.route.id;
-    if (this.lastLayoutConfig?.id === id && this.lastLayoutConfig?.display === display) return;
-    this.lastLayoutConfig = { id, display };
-    this.layoutService.configure({ contentDisplayType: display });
   }
 
 }
